fix: if Date.now returns a date get its timestamp
